move the program clause stuff into its own module
